Moved model classes to the View package, Added logged user type control, Navigation bar shows the field related with the user type
