--- app/src/main/java/com/douglas/thecarserviceapp/ViewAppointments.java
@@ -1,44 +1,0 @@
-package com.douglas.thecarserviceapp;
-
-import androidx.appcompat.app.AppCompatActivity;
-import androidx.core.view.GravityCompat;
-import androidx.drawerlayout.widget.DrawerLayout;
-import androidx.recyclerview.widget.LinearLayoutManager;
-import androidx.recyclerview.widget.RecyclerView;
-
-import android.os.Bundle;
-import android.view.View;
-import android.widget.ImageView;
-
-public class ViewAppointments extends AppCompatActivity {
-    DrawerLayout drawerLayout;
-    ImageView btMenu;
-    RecyclerView recyclerView;
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_view_appointments);
-
-        drawerLayout = findViewById(R.id.drawer_layout);
-        btMenu = findViewById(R.id.bt_menu);
-        recyclerView = findViewById(R.id.recycler_view);
-
-        recyclerView.setLayoutManager(new LinearLayoutManager(this));
-        recyclerView.setAdapter(new MainAdapter(this, BookAnAppointment.arrayList));
-
-        btMenu.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                drawerLayout.openDrawer(GravityCompat.START);
-            }
-        });
-    }
-
-    @Override
-    protected void onPause() {
-        super.onPause();
-        //close drawer
-        BookAnAppointment.closeDrawer(drawerLayout);
-    }
-}
