+++ app/src/main/java/com/douglas/thecarserviceapp/view/BookAnAppointment.java
@@ -1,0 +1,74 @@
+package com.douglas.thecarserviceapp.view;
+
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.core.view.GravityCompat;
+import androidx.drawerlayout.widget.DrawerLayout;
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
+
+import android.os.Bundle;
+import android.view.View;
+import android.widget.ImageView;
+
+import com.douglas.thecarserviceapp.R;
+import com.douglas.thecarserviceapp.adapter.MainAdapter;
+import com.douglas.thecarserviceapp.app.AppManager;
+
+import java.util.ArrayList;
+
+public class BookAnAppointment extends AppCompatActivity {
+    //variable
+    DrawerLayout drawerLayout;
+    ImageView btMenu;
+    RecyclerView recyclerView;
+    static ArrayList<String> arrayList = new ArrayList<>();
+    MainAdapter adapter;
+
+    public static void closeDrawer(DrawerLayout drawerLayout) {
+        if(drawerLayout.isDrawerOpen(GravityCompat.START)){
+            drawerLayout.closeDrawer(GravityCompat.START);
+        }
+    }
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_bookanappointment);
+
+        drawerLayout = findViewById(R.id.drawer_layout);
+        btMenu = findViewById(R.id.bt_menu);
+        recyclerView = findViewById(R.id.recycler_view);
+
+        arrayList.clear();
+        if(AppManager.instance.user.isCustomer()){
+            arrayList.add("Book an Appointment");
+            arrayList.add("Search Service Provider");
+            arrayList.add("View Appointments");
+            arrayList.add("Service History");
+            arrayList.add("Profile");
+            arrayList.add("Logout");
+        }else if (AppManager.instance.user.isProvider()){
+            arrayList.add("View Appointments");
+            arrayList.add("Create Customer");
+            arrayList.add("Search Customers");
+            arrayList.add("Service History");
+            arrayList.add("Logout");
+        }
+        adapter = new MainAdapter(this, arrayList);
+        recyclerView.setLayoutManager(new LinearLayoutManager(this));
+        recyclerView.setAdapter(adapter);
+
+        btMenu.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View view) {
+                drawerLayout.openDrawer(GravityCompat.START);
+            }
+        });
+    }
+
+    @Override
+    protected void onPause() {
+        super.onPause();
+        closeDrawer(drawerLayout);
+    }
+}
