+++ app/src/main/res/layout/activity_bookanappointment.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:id="@+id/drawer_layout"
-    tools:context=".BookAnAppointment">
+    tools:context=".view.BookAnAppointment">
 
     <LinearLayout
         android:layout_width="match_parent"
