+++ app/src/main/res/layout/activity_profile.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:id="@+id/drawer_layout"
-    tools:context=".Profile">
+    tools:context=".view.Profile">
 
     <LinearLayout
         android:layout_width="match_parent"
