+++ app/src/main/res/layout/activity_service_history.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:id="@+id/drawer_layout"
-    tools:context=".ServiceHistory">
+    tools:context=".view.ServiceHistory">
 
     <LinearLayout
         android:layout_width="match_parent"
