+++ app/src/main/res/layout/activity_view_appointments.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:id="@+id/drawer_layout"
-    tools:context=".ViewAppointments">
+    tools:context=".view.ViewAppointments">
 
     <LinearLayout
         android:layout_width="match_parent"
